+++ config/routes.rb
@@ -66,15 +66,6 @@ Rails.application.routes.draw do
 
   resources :broadcasts, path: "editor/blog"
 
-  resources :broadcast_comments do
-    collection do
-      post :preview
-    end
-    member do
-      post :vote
-    end
-  end
-
   resources :images do
     collection do
       post :upload, action: :create_multiple