+++ db/migrate/20180126201332_remove_broadcast_comment_id_from_notifications.rb
@@ -1,0 +1,6 @@
+class RemoveBroadcastCommentIdFromNotifications < ActiveRecord::Migration[5.2]
+  def change
+    remove_index :notifications, :broadcast_comment_id
+    remove_column :notifications, :broadcast_comment_id, :integer
+  end
+end